\name{GetNgramIdx}
\alias{GetNgramIdx}
\title{GetNgramIdx}
\usage{
  GetNgramIdx(str, n)
}
\arguments{
  \item{str}{A character vector whose components are
  strings containing one or more words}

  \item{n}{The length of the n-grams for which vector
  indices are to be returned.}
}
\description{
  GetNgramIdx
}

