\docType{package}
\name{leghist-package}
\alias{leghist-package}
\title{Mapping of textual evolution of legislation}
\description{
  leghist provides functionality within the R programming
  language to automate the textual evolution of legislation
  from its introduction as a bill, through the amendment
  process, until finalization.
}
\details{
  leghist provides five core pieces of functionality: 1.
  Mapping of sections between editions of bills 2. Mapping
  of amendments to their location in bills 3.
  Identification of discarded material 4. Modeling of the
  content of added and discarded material 5. Visualization
  of the flow of bill content, by subject matter area, from
  the initial to final bill.
}
\keyword{package}

