\name{MakeShapesPAS}
\alias{MakeShapesPAS}
\title{MakeShapesPAS}
\usage{
  MakeShapesPAS(a, f, rejected.shape)
}
\arguments{
  \item{a}{the number of amendments.}

  \item{f}{the number of paragraphs in the final bill.}

  \item{rejected.shape}{the shape of the rejected node.}
}
\value{
  A vector of shapes for a PAS graph. author Hillary
  Sanders
}
\description{
  Creates a vector of node (vertex) shapes for a
  PlotAmendsSuccess() graph.
}

