\name{ValidateSelection}
\alias{ValidateSelection}
\title{ValidateSelection}
\usage{
  ValidateSelection(selection, valid.matches)
}
\arguments{
  \item{selection}{the user input from the command line}

  \item{valid.matches}{a set of valid user inputs}
}
\value{
  TRUE if the input is valid, else FALSE
}
\description{
  Validates the user input for the encoder routine and
  returns a boolean
}
\author{
  Mark Huberty
}

