\name{VertexSizesPAS}
\alias{VertexSizesPAS}
\title{VertexSizesPAS}
\usage{
  VertexSizesPAS(a, f, rejected.scale)
}
\arguments{
  \item{a}{the number of amendments.}

  \item{f}{the number of paragraphs in the final bill.}

  \item{rejected.shape}{the shape of the rejected node.}
}
\value{
  A vector of sizes for a PAS graph. author Hillary Sanders
}
\description{
  Creates a vector of node (vertex) sizes for a
  PlotAmendsSuccess() graph.
}

